Extract ModelStub to remove duplication
